Keep msg buffer around to send to siblings
For the fed_mgr, instead of packing up the job_desc again for will_runs
and allocations to each sibling just send the buffer that came in.

Since job_allocate will modify the received job_desc_msg_t, this also
makes it easy to make a copy of the received job_desc_msg_t to do a
willrun on the receiving cluster before allocating resources.